- Moved more parsed datatypes away from trivial types

+++ java/src/tv/porst/swfretools/parser/tags/SoundStreamHeadTag.java
@@ -1,5 +1,6 @@
 package tv.porst.swfretools.parser.tags;
 
+import tv.porst.splib.io.INT16;
 import tv.porst.splib.io.UINT16;
 import tv.porst.swfretools.parser.structures.RecordHeader;
 
@@ -9,7 +10,7 @@ public class SoundStreamHeadTag extends Tag {
 			final int playbackSoundRate, final int playbackSoundSize,
 			final int playbackSoundType, final int streamSoundCompression,
 			final int streamSoundRate, final int streamSoundSize, final int streamSoundType,
-			final UINT16 streamSoundSampleCount, final int latencySeek) {
+			final UINT16 streamSoundSampleCount, final INT16 latencySeek) {
 		super(header);
 	}
